+++ src/v1/BasicTypes/Strings/index.ts
@@ -36,6 +36,7 @@ export * from "./isString";
 export * from "./mustBeString";
 export * from "./validateString";
 export * from "./validateStringLengthBetween";
+export * from "./validateStringMinLength";
 export * from "./validateStringMatches";
 export * from "./validateStringStartsWith";
 export * from "./validateStringValue";